Before to destroy the spash-screen, we need to check if it is Void.
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@15321 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/library/vision2/implementation/mswin/kernel/ev_application_imp.e
@@ -176,7 +176,7 @@ feature {NONE} -- WEL Implemenation
 		do
 			create controls_dll.make_with_flags (Icc_userex_classes
 					+ Icc_win95_classes + Icc_cool_classes
-					+ Icc_bar_classes)
+					+ Icc_bar_classes + Icc_updown_class)
 			create rich_edit_dll.make
 		end
 
@@ -193,8 +193,10 @@ feature {NONE} -- Message loop, we redefine it because the user
 			dlg: POINTER
 		do
 			-- Destroy the spash screen
-			splash_screen.destroy
-			splash_screen := Void
+			if splash_screen /= Void then
+				splash_screen.destroy
+				splash_screen := Void
+			end
 
 			-- `accel' and `main_w' are declared
 			-- locally to get a faster access.
